Use generic module-level address decoder
This in preparation for adding more devices on the memory bus (UART, SDRAM).

Also increase the duration of initial reset; this seemed to be necessary to
make things work reliably.

Also lower the clock to 24 MHz. This was not seen to be necessary yet, but
icetime output said 40 MHz at one point, and I do not want to suffer under
marginal timing during development.